weather_api/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2==2.9.9
python-dotenv==1.0.0
requests==2.31.0
spacy==3.7.2
